Add files via upload
The main focus of the changes made today was to move all of the variables that the user would have to specify out of the body of the choropleth() function. These variables are now function arguments which are to be specified in the function call. This means that the script is now much more reusable. To output a choropleth using another data set, or to output a choropleth with the colour weighting based on another column, you simply have to alter the input arguments rather than changing the body of the function. I have ran this code and saved the output under 'choropleth_map_sample_data_SumAssured.png' - which can be viewed in this repository.