Add files via upload
After analysing the outputs from the code in geopandas_project_1.py it became clear that the outputs were not accurately representing the data when you opt for a market value (MV) choropleth output. I realised that our final geodataframe had several rows for each LAD which causing the colours to not representing the total market value of the policies held in each district. 

To address this I decided to find a way to aggregate the market value for each postcode. Then I achieved an aggregate market value of the policies for each district. Then I combined this dataframe to the 'policy_count_dataframe' in geopandas_project_1.py. 

Thus, this allowed me to achieve a final geodataframe with one row per LAD, containing a single aggregated market value of the policies and a single value of the number of policies. The code in this file is much more simpler and concise as now it only has one geodataframe meaning that much of the logic is no longer dependent on the choice of weighted variable.