Add files via upload
Some extra coding was needed to overcome an issue of LAD's not being plotted if they didn't have a postcode in the client data. So, I added rows to our final geodataframe of these LAD's and gave them 0 values for their market value and policy count. The order of the code was rearranged slightly to achieve this. 
I realised that there was still a couple of parameters which needed to be defined to make the code completely reusable for different data sets - avoiding the need of the user to change the body of the function for each data set. These changes were made on the final date of my summer internship at OAC.